+++ hack/generate-kustomize-patches.sh
@@ -20,21 +20,13 @@ KUSTOMIZE_INPUT_DIR="$ROOT_DIR/config/helm/input"
 
 # Download upstream manifests
 helm_values="$HELM_DIR/values.yaml"
-# Giant Swarm specific, since we don't use GitHub releases in https://github.com/giantswarm/cluster-api
-version="$(yq e -e '.images.tag' "$helm_values")" || { echo "Could not find image tag value"; exit 1; }
+org="kubernetes-sigs"
+repo="cluster-api"
+version="$(yq e '.images.tag' "$helm_values")"
 release_asset_filename="cluster-api-components.yaml"
+url="https://github.com/$org/$repo/releases/download/$version/${release_asset_filename}"
 mkdir -p "$KUSTOMIZE_INPUT_DIR"
-# Image does not have a shell or `cat` installed, so extract the file using busybox
-empty_context="$(mktemp -d)"
-cat >"${empty_context}/Dockerfile.manifest" <<EOF
-FROM docker.io/giantswarm/cluster-api-controller:${version} as src
-FROM docker.io/library/busybox:1
-COPY --from=src /for-cluster-api-app-only/cluster-api-components.yaml /file
-EOF
-docker build -f "${empty_context}/Dockerfile.manifest" -t manifest "${empty_context}"
-rm -r "${empty_context}"
-docker run --rm manifest cat /file >"$KUSTOMIZE_INPUT_DIR/$release_asset_filename"
-[ "$(grep -c ^ "$KUSTOMIZE_INPUT_DIR/$release_asset_filename")" -gt 20000 ] || { >&2 echo "Downloaded ${release_asset_filename} does not seem right"; exit 1; }
+curl -L "$url" -o "$KUSTOMIZE_INPUT_DIR/${release_asset_filename}"
 
 # Update kustomize patches for webhooks. We do this for every CRD